+++ src/package-json/packageJson.ts
@@ -1,12 +1,18 @@
 import {dirnameFromMeta, readJsonFile, searchFileInDirectoryTree} from "#core/FS/utils";
+import {isSea} from 'node:sea'
+
+declare const __APP_VERSION__: string
 
 export type PackageJson = {
   version: string
   name: string
 }
-export const readPackageJson = (): PackageJson | undefined => {
-  const jsonPath = searchFileInDirectoryTree('package.json', dirnameFromMeta(import.meta))
 
+export const readPackageJson = (): PackageJson | undefined => {
+  if (isSea()) {
+    return {name: 'tunli', version: __APP_VERSION__}
+  }
+  const jsonPath = searchFileInDirectoryTree('package.json', dirnameFromMeta(import.meta))
   if (!jsonPath) return
   return readJsonFile(jsonPath) as PackageJson
 }